Remove squashfs from netboot
The filesystem is read exactly once on boot, and the entire contents
is copied to the actual root filesystem. There's no need for it to
remain mounted, or even be a filesystem at all! Replace it with a
single xz tarball that's extracted on boot.

Also includes tar-store.nix from a795af32106eee3eba5b18a7723d0118c796c521