upgrade from NixOS 19.09 -> 20.09
I was able to build the x86_64-linux images locally:
  $ NIX_PATH=nixpkgs=channel:nixos-20.09 nix-build instances/x86_64-linux.nix

Unfortunately I can't test the resulting image in QEMU as the initrd
size is not supported by QEMU.

I had to remove the import of the specialization.nix as those options
were already present in NixOS 20.09 [1]. It appears to be Graham that pushed
those changes upstream? I am not sure.

[1] https://github.com/NixOS/nixpkgs/commit/ec2d28e3233c4c04fe6afe13e828691d099167ae